Modify Dockerfile template to preserve the 64-bit Wine binary for 32-bit builds (#12)
Currently when building a Wine container with 32-bit support, the Wine binary
gets clobbered by the 32-bit version. This change fixes that behaviour and
instead renames the 64-bit binary to 'wine64' for 32-bit builds. In the case of
a 64-bit only build, 'wine64' is created as a symlink to 'wine' instead.

Co-authored-by: Aiden De Loryn <aiden-deloryn@users.noreply.github.com>