Modify Dcokerfile template to flatten directory structure for Wine PDBs and DLLs
Currently the entire directory structure is preserved for all Wine DLLs and PDBs. This change flattens the structure so that all DLL's are stored in a single folder alongside most of the PDB files. Any duplicate PDB files are stored in subdirectories in order to prevent loss of data. 32-bit and 64-bit files are stored in their own separate subfolders.